+++ src/storage/column_vector/value.cpp
@@ -33,6 +33,71 @@ import type_info;
 
 namespace infinity {
 
+namespace {
+template <typename T>
+void Embedding2JsonInternal(const EmbeddingType &embedding, size_t dimension, nlohmann::json& embedding_json) {
+    for (size_t i = 0; i < dimension; ++i) {
+        embedding_json.push_back(((T *)(embedding.ptr))[i]);
+    }
+}
+
+void BitmapEmbedding2JsonInternal(const EmbeddingType &embedding, size_t dimension, nlohmann::json& embedding_json) {
+    if(dimension % 8 != 0) {
+        Status status = Status::SyntaxError("Binary embedding dimension should be the times of 8.");
+        LOG_ERROR(status.message());
+        RecoverableError(status);
+    }
+
+    auto *array = reinterpret_cast<const u8*>(embedding.ptr);
+    for (size_t i = 0; i < dimension / 8; ++i) {
+        const u8 byte = array[i];
+        for (size_t j = 0; j < 8; ++j) {
+            int8_t elem = (byte & (1 << j)) ? 1 : 0;
+            embedding_json.push_back(elem);
+        }
+    }
+}
+
+void Embedding2Json(const EmbeddingType &embedding, EmbeddingDataType type, size_t dimension, nlohmann::json& embedding_json) {
+    switch (type) {
+        case kElemBit: {
+            BitmapEmbedding2JsonInternal(embedding, dimension, embedding_json);
+            break;
+        }
+        case kElemInt8: {
+            Embedding2JsonInternal<int8_t>(embedding, dimension, embedding_json);
+            break;
+        }
+        case kElemInt16: {
+            Embedding2JsonInternal<int16_t>(embedding, dimension, embedding_json);
+            break;
+        }
+        case kElemInt32: {
+            Embedding2JsonInternal<int32_t>(embedding, dimension, embedding_json);
+            break;
+        }
+        case kElemInt64: {
+            Embedding2JsonInternal<int64_t>(embedding, dimension, embedding_json);
+            break;
+        }
+        case kElemFloat: {
+            Embedding2JsonInternal<float>(embedding, dimension, embedding_json);
+            break;
+        }
+        case kElemDouble: {
+            Embedding2JsonInternal<double>(embedding, dimension, embedding_json);
+            break;
+        }
+        default: {
+            Status status = Status::SyntaxError("Unexpected embedding type");
+            LOG_ERROR(status.message());
+            RecoverableError(status);
+        }
+    }
+}
+
+}
+
 // Value maker
 Value Value::MakeValue(DataType type) {
     Value value(type.type(), type.type_info());
@@ -1071,7 +1136,7 @@ void Value::AppendToJson(const String &name, nlohmann::json &json) {
                 UnrecoverableError(error_message);
             }
             const EmbeddingT embedding(const_cast<char *>(data_span.data()), false);
-            EmbeddingT::Embedding2Json(embedding, embedding_info->Type(), embedding_info->Dimension(), json[name]);
+            Embedding2Json(embedding, embedding_info->Type(), embedding_info->Dimension(), json[name]);
             return;
         }
         case LogicalType::kTensor: {
